fix: Update PokemonDetails model to remove unnecessary type conversion; enhance DetailPage with loading indicator and display of Pokemon details

+++ lib/models/pokemon_details.model.dart
@@ -24,8 +24,8 @@ class PokemonDetails {
     return PokemonDetails(
       name: data['name'],
       id: data['id'],
-      height: data['height'].toDouble(),
-      weight: data['weight'].toDouble(),
+      height: data['height'],
+      weight: data['weight'],
       baseExperience: data['base_experience'],
       stats: (data['stats'] != null ? data['stats'] as List<dynamic> : null)
           ?.map((stat) => Stat.fromJson(stat))
